fix : 상품 주문 시 상품 서비스로 상품 수량이 전파되지 않던 문제 해결

+++ product-service/src/main/java/com/productservice/mapper/MapToStockUpdateDtoMapper.java
@@ -8,7 +8,7 @@ public class MapToStockUpdateDtoMapper {
     public static StockUpdateDto map(Map<String, Object> map) {
         return StockUpdateDto.builder()
                 .productId((String) map.get("productId"))
-                .quantity((Long) map.get("quantity"))
+                .quantity(((Integer) map.get("quantity")).longValue())
                 .build();
     }
 }
